+++ lib/src/rules/max_length_rule.dart
@@ -1,4 +1,4 @@
-import 'package:flrx_validator/rule.dart';
+import 'package:flrx_validator/src/rules/rule.dart';
 
 /// A [Rule] subclass validating if the length of input is less than [maxLength].
 class MaxLengthRule extends Rule<String> {